% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mln_helpers.R
\name{alr}
\alias{alr}
\title{alr: Additive Log Ratio calculation}
\usage{
alr(P)
}
\arguments{
\item{P}{A vector or matrix (rows) of probabilities}
}
\value{
A vector or matrix (rows) of log-odds
}
\description{
Estimating underlying log-odds given probability vector (last category is treated as baseline)
}
\examples{
\dontrun{
# Example usage:
alr(c(.1,.3,.6))
}

}
