% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_data_examples.R
\name{clean_Lahman_data}
\alias{clean_Lahman_data}
\title{Prepare Cleaned Lahman Baseball Data for Multinomial Modeling}
\usage{
clean_Lahman_data()
}
\value{
A list with components:
\describe{
\item{Y}{An \eqn{N \times J} matrix of batting outcome counts (default: HR, BB, SO, and Other).}
\item{X}{An \eqn{N \times p} numeric matrix of standardized predictors (default: p = 10).}
\item{Z}{An \eqn{N \times m} indicator matrix for random intercepts by playerID.}
}
}
\description{
This function cleans and processes data from the \pkg{Lahman} database to produce a usable dataset
for modeling MLB player batting outcomes. It returns multinomial response data (\code{Y}), standardized predictors (\code{X}),
and a random intercept design matrix (\code{Z}) for player-level effects.
}
\details{
The output includes only non-pitcher players with PA > 200 from the American and National Leagues
(post-1960; expansion era), and aggregates data by player-season (grouping across stints and teams).
It collapses batting outcomes into home runs, walks, strikeouts, and other outcomes.
Predictors include physical measurements, age (and age^2), batting side, league, and a grouped fielding position indicator.
Investigating the code, it should not be difficult to adjust the outcomes investigated and/or the predictors.
}
\examples{
\dontrun{
data <- clean_Lahman_data()
str(data$X)
colnames(data$Z)[1:5]
}

}
