% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mln_helpers.R
\name{compress_counts}
\alias{compress_counts}
\title{compress_counts: Smithson-Verkuilen Correction}
\usage{
compress_counts(w)
}
\arguments{
\item{w}{A vector or matrix (rows) of multinomial counts, to be compressed in the presence of zero counts}
}
\value{
A vector or matrix (rows) of compressed multinomial counts
}
\description{
This function implements the correction from Smithson and Verkuilen (2006) to handle zero counts. Input is a vector or matrix of multinomial counts to be compressed row-wise
}
\examples{
\dontrun{
# Example Usage:
# in the presence of zero counts
compress_counts(c(1,0,3))
# if no zero counts
compress_counts(c(1,2,3))
}

}
