% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mln_helpers.R
\name{dmnl_loglik}
\alias{dmnl_loglik}
\title{dmnl_loglik: Multinomial Logistic-Normal Log Likelihood}
\usage{
dmnl_loglik(Y, W)
}
\arguments{
\item{Y}{A vector or matrix (rows) of the latent log-odds for observations (d columns).}

\item{W}{A vector or matrix (rows) of multinomial counts (d+1 columns).}
}
\value{
The log likelihood value (numeric).
}
\description{
Calculate the log likelihood for the Multinomial Logistic-Normal distribution given latent Y and counts W.
}
