% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MMLN-package}
\alias{MMLN-package}
\title{MMLN: Mixed Effects Multinomial Logistic Normal Models}
\description{
Tools for fitting and evaluating mixed effects multinomial regression models, including methods for calculating squared Mahalanobis distance residuals and testing model fit.
}
\author{
\strong{Maintainer}: Eric Gerber \email{e.gerber@northeastern.edu} (\href{https://orcid.org/0009-0008-0042-7051}{ORCID})

}
