% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_res.R
\name{plot_trace_and_summary}
\alias{plot_trace_and_summary}
\title{plot_trace_and_summary: Traceplots and summary tables for MCMC chains}
\usage{
plot_trace_and_summary(chain_array, param_name = "param", max_rows = 4)
}
\arguments{
\item{chain_array}{An array (rows by cols by iterations) obtained via simplify2array(fit$chain) from a model fit.}

\item{param_name}{Character. Name of the parameter to label the plots (default "param").}

\item{max_rows}{Integer. Maximum number of rows of parameter indices to plot (default 4).}
}
\value{
A 3D array (cols by stats by rows) of summary statistics: mean, sd, 2.5\%, 97.5\% quantiles.
}
\description{
This function plots and provides posterior means and credible intervals for the MCMC samples of fitted MLN models.
}
\examples{
\dontrun{
# Example usage:
sim <- simulate_mixed_mln_data(
m       = 30,
n_i     = 10,
p       = 2,
d       = 2,
beta    = matrix(c(-2, 0.5, 1, -0.), nrow = 2, ncol = 2),
Sigma   = diag(c(1.2, .7)),
Phi     = diag(c(0.5, 0.3)),
PA_mean = 100)

res <- MMLN(
W       = sim$W,
X       = sim$X,
Z       = sim$Z,
n_iter  = 1000,
burn_in = 180,
thin    = 2,
proposal= "normbeta",
verbose = TRUE)

beta_chain_array <- simplify2array(res$beta_chain)
plot_trace_and_summary(beta_chain_array, "beta")

}

}
