% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mln_helpers.R
\name{pstartoy}
\alias{pstartoy}
\title{pstartoy: Logit Transformation}
\usage{
pstartoy(pstarvec)
}
\arguments{
\item{pstarvec}{A numeric vector of probabilities (between 0 and 1).}
}
\value{
A numeric vector of log-odds values.
}
\description{
Compute log-odds (logit) from probability values.
}
\examples{
\dontrun{
pstartoy(c(0.2, 0.5, 0.8))
}

}
