% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_res.R
\name{summary.mdres}
\alias{summary.mdres}
\title{summary.mdres: Summary and diagnostics for Mahalanobis residuals}
\usage{
\method{summary}{mdres}(object, ...)
}
\arguments{
\item{object}{Numeric vector of residuals (class 'mdres').}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the Kolmogorov-Smirnov test object.
}
\description{
Perform a Kolmogorov-Smirnov test and generate a QQ plot for Mahalanobis residuals.
}
