% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mln_helpers.R
\name{ytopstar}
\alias{ytopstar}
\title{ytopstar: Inverse Logit Transformation}
\usage{
ytopstar(yvec)
}
\arguments{
\item{yvec}{A numeric vector of log-odds values.}
}
\value{
A numeric vector of probabilities (between 0 and 1).
}
\description{
Compute probabilities from log-odds values.
}
\examples{
\dontrun{
ytopstar(c(-1, 0, 1))
}

}
