% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_res.R
\name{compute_dic}
\alias{compute_dic}
\title{compute_dic: Deviance Information Criterion}
\usage{
compute_dic(loglik_chain, loglik_hat)
}
\arguments{
\item{loglik_chain}{Numeric vector of posterior log-likelihood samples across iterations.}

\item{loglik_hat}{Numeric. The posterior mean of the log-likelihood (evaluated at posterior means).}
}
\value{
A list with elements:
\describe{
\item{DIC}{The Deviance Information Criterion value.}
\item{pD}{The effective number of parameters.}
}
}
\description{
Computes the DIC (Deviance Information Criterion) and effective number of parameters pD.
}
\examples{
\dontrun{
# Example usage:
sim <- simulate_mixed_mln_data(
m       = 30,
n_i     = 10,
p       = 2,
d       = 2,
beta    = matrix(c(-2, 0.5, 1, -0.), nrow = 2, ncol = 2),
Sigma   = diag(c(1.2, .7)),
Phi     = diag(c(0.5, 0.3)),
n_mean = 100)

res <- MMLN(
Y       = sim$Y,
X       = sim$X,
Z       = sim$Z,
n_iter  = 1000,
burn_in = 180,
thin    = 2,
proposal= "normbeta",
verbose = TRUE)

ll_chain <- lapply(seq_len(length(res$w_chain)),
function(i){
dmnl_loglik(Y = res$y_chain[[i]], Y = sim$Y)})

W_hat <- alr(compress_counts(sim$Y)/rowSums(sim$Y))
ll_hat <- dmnl_loglik(W = W_hat, Y = sim$Y)

compute_dic(ll_chain, ll_hat)

}

}
