% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mln_helpers.R
\name{dmnl_loglik}
\alias{dmnl_loglik}
\title{dmnl_loglik: Multinomial Logistic-Normal Log Likelihood}
\usage{
dmnl_loglik(W, Y)
}
\arguments{
\item{W}{A vector or matrix (rows) of the latent log-odds for observations (J-1 columns).}

\item{Y}{A vector or matrix (rows) of multinomial counts (J columns).}
}
\value{
The log likelihood value (numeric).
}
\description{
Calculate the log likelihood for the Multinomial Logistic-Normal distribution given latent W and counts Y.
}
